Update patch set 2
Patch Set 2:

Yeah the evil DS_Store is needed. But for the rest of it i m not sure. Maybe the cocoa rules should be added later if we were to use it. All the rest seems to have been moved elsewhere (DerivedData is now in ~/Library/Developer/Xcode/DerivedData/, Build is in DerivedData now)

Patch-set: 2
Label: Code-Review=0