Update patch set 3
Patch Set 3:

What about Cocoapods ?

I think we should not try to add a minimal .gitignore and amend it along the development but start with a template covering everything. I think it is more error-proof.

A gitignore is simply a text file, having 100 lines instead of 10 is not that bad.

Patch-set: 3
Label: Code-Review=0