Update patch set 3
Patch Set 3: -Code-Review

> > No this is related because it fix the commit rattached to this
 > > ticket, which shouldnt have been merged like this in the first
 > > place.
 > 
 > Nevermind i thought this was the daemon patch. hum well maybe a
 > ticket but seems a little overkill for such a trivial change

Oh ok no problem. I think that it is interesting to have one to log what you do, why you do what you do, and finally things you want to say about that.
For example, I understood that this patch allows Ring to build and run but in simulator only. It could be useful to log this in the commit message and/or in an associated Tuleap.
It is nothing but a suggestion.

Patch-set: 3