daemonservice: introduce service
This patch adds the deamon service to the project.

This service will handle the main capabilities of the Ring Daemon,
such as starting and stopping it.

For the moment, there is no callback, notifications, or observers to
that service. This let us the flexibility to implement it with
whatever we want in the future (Rx-like, NSNotifications etc...).

Tuleap: #1378
Change-Id: Ib560977f9fd1dacd45e2afc63613870a3d60630f