account service: modify bridging
This patch reorganizes the structure of the bridging between the Swift
client and the Daemon.

It isolates the code related to the account service. That's why there
is a renaming of the ConfigurationManagerAdaptor.
We want it to be specialized in the what the account service can do.

The files meant to make the bridge between the client and the daemon
are now stored in a dedicated file in the project structure.

The broadcast by NSNotification posts is still used here.

Tuleap: #1405
Change-Id: I2d8f0d361ed9394c98a30b0703ad3d49b18cde4a