dringadaptor: use main thread for calls
This patchs ensures that the dringadaptor uses the application's main
thread to make the calls to the daemon.

This is necessary because of the conception of the daemon.

It captures and registers the thread that initializes and starts it
and considers it as its main thread.
pjsip then also uses it. It does not tolerate any other threads than
the ones that it had already registered.

Using the app main thread guarantees that is remains the same during
the whole lifecyle.

Tuleap: #1423
Change-Id: Id03933f3f1a4c0428899af3593b35fc92a873e34