wizard: bind V to VM for create Ring account
This patch creates a basic binding between the view and the modelview
of the add account screen.

It is done by using RxSwift.

For the moment, the unique button of the screen is bound. The action
between the VM and the AccountService is still strongly link. This
will be changed in a next patch to use Rx instead.

Change-Id: Iebc4f49edd0de4a0ec1db70af854d0d6f98f336e
Tuleap: #1400