accountModel: create model stack
This patch creates the stack of models that the AccountModel will need
to exist.

It also makes a little refactor to make the AccountModel very simple:
variables and init methods.
Some of these elements that have been moved are now gathered in a
dedicated helper that exposes some getters and setters.

Finally, some code is now commented in the high layers to allow code
compilation. This code refers to parts that are not designed or
written yet.

Tuleap: #1542
Change-Id: Ie52075c9f06d6efaa824caac457d30b859c17658