Update patch set 2
Patch Set 2: Code-Review-1

I think that the constraints you set are not correct, but let my know if I'm wrong, it must depend on what you want to do.

For example, ringImageView, has its constraints and is scaleToFill. Furthermore, the height of the imageView changes if the text of the descriptionLabel changes.

welcomeLabel can not go 2 lines if the text is too long.

Also, the buttons can be very large on a large screen (32 left and right).

Patch-set: 2
Label: Code-Review=-1