CreateRingAccount: Add Rx+MVVM bindings for the form validation
Add some bindings from the CreateRingAccountViewController to the
CreateRingAccountViewModel to validate the following user inputs :

- username (optional)
- password (must be 6 characters min.)
- repeat password (must be equal to the password)

If conditions are not verified. The Create Account button is
disabled.
Username TextField has a 0.5s time interval between user inputs.

Bindings are made with Variabes and Observers into the ViewModel.
The username validation is not complete yet. So there is a 2s timer
to simulate the request.

Tuleap: #1400
Change-Id: Ib03797e365972aefafbc75c5a8b798e8ba659cfc