Messages: Add messages screen to read messages and reply
Add MVVM, services and adpater to receive and reply to messages

This allows to read received messages from other users and reply to
them.

The messages screen uses a TableView that scroll automatically to the
bottom if the user a new message is received. Messages bubbles are
made with simple cells with a label that move to the right or the
left.

Change-Id: I1b6bd6fd36b88daab9d39cb5745d1ed953868ead