Smartlist: Add a search bar to filter conversations and lookup users
Add the search bar on the top of the smartlist and the list of found
results.

Add an other list for search results. This list is splited in 2
sections for filered and found results.

Add "O Conversations" label if there is no conversations

If the contact is not found locally, a new conversation is created

Change-Id: I7985f0644f97063875bfb02159d44dd1aa8b731e