Smartlist: Add a search bar to filter conversations and lookup users
Add Contact service to lookup users from user name or RingId.
Add the search bar on the top of the smartlist and the list of found
results.

If the contact is not found locally, the Contact service create a new
conversation.

The Contact service keep a local list of not trusted users.

Change-Id: I7985f0644f97063875bfb02159d44dd1aa8b731e