accountsService: introduce new AccountsService
This new accountsService is designed to avoid any model retention /
duplication from the daemon.
The daemon is used as a repository.

It now takes the delegation from the accountAdapter, making the old
accountAdapter not responding to the daemon events.

This new service will be used more and more to finally replace the old
one as soon as possible. For now, the two coexist.

Change-Id: I5a925312254c1b2df1c51966a1282da7850a075d
Tuleap: #1752