Update patch set 5
Patch Set 5: Code-Review-1

Could this new logic be done in the VCardUtils class? 

In my understanding, and it might of course be wrong, this class should handle once and for all the logic behind how to get the vcard. This intelligence would then be located at one place.

What do you think?

Patch-set: 5
Label: Code-Review=-1