vcard: fix parsing
This patch removes CNContactVCardSerialization for
vCards and replaces it with custom vCard parsing.
This change was made because CNContactVCardSerialization
requires a providing charset to properly parse special
characters that may not be present in vCards created on
other platforms. Additionally, CNContacts were not being
utilized, so removing this step streamlines the parsing
of contacts.

Gitlab: #267
Change-Id: I9207b8780e5893a90691e060eb52c985a2996058